# docs/openapi.yaml
openapi: 3.0.3
info:
  title: KubeOrch Core API
  description: REST API for the KubeOrch Kubernetes workflow orchestration platform.
  version: 0.0.3
  x-stability-level: beta
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: KubeOrch Community
    url: https://github.com/KubeOrch
...
servers:
  - url: http://localhost:3000/v1
    description: Local development
  - url: https://api.kubeorch.dev/v1
    description: Production
...
paths:
...
  /api/workspaces/{workspaceId}/artifacts:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Artifacts]
      summary: List registered artifacts
      description: Returns immutable artifact evidence in descending creation order without fetching evidence references.
      operationId: listArtifacts
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [artifacts:read]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Workspace artifacts
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtifactList"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Artifacts]
      summary: Register immutable artifact evidence
      description: Registers a digest-pinned OCI image and metadata. This operation never fetches evidence, builds source, or deploys to a cluster.
      operationId: createArtifact
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [artifacts:write]
      x-kubeorch-idempotency: required
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateArtifactRequest"
      responses:
        "201":
          description: Artifact registered, or an existing identical artifact replayed
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "413":
          $ref: "#/components/responses/APIError"
  /api/workspaces/{workspaceId}/artifacts/{artifactId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ArtifactId"
    get:
      tags: [Artifacts]
      summary: Get registered artifact evidence
      operationId: getArtifact
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [artifacts:read]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Immutable artifact evidence
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artifact"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
  /api/workspaces/{workspaceId}/applications/{applicationId}/releases:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ApplicationId"
    get:
      tags: [Releases]
      summary: List application releases
      description: Returns immutable releases in descending creation order. Listing does not inspect a cluster.
      operationId: listReleases
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [releases:read]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageCursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Releases for the route application
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseList"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
    post:
      tags: [Releases]
      summary: Create an immutable application release
      description: Binds one application revision to registered workspace artifacts. It records a handoff only and never mutates workflows or clusters.
      operationId: createRelease
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [releases:write]
      x-kubeorch-idempotency: required
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseRequest"
      responses:
        "201":
          description: Release created, or the original creation response replayed
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
            Idempotency-Replayed:
              $ref: "#/components/headers/IdempotencyReplayed"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "400":
          $ref: "#/components/responses/APIError"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
        "409":
          $ref: "#/components/responses/APIError"
        "413":
          $ref: "#/components/responses/APIError"
  /api/workspaces/{workspaceId}/applications/{applicationId}/releases/{releaseId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/ApplicationId"
      - $ref: "#/components/parameters/ReleaseId"
    get:
      tags: [Releases]
      summary: Get an application release
      operationId: getRelease
      x-stability-level: beta
      x-kubeorch-workspace-boundary: workspace
      x-kubeorch-required-scopes: [releases:read]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Immutable application release
          headers:
            X-Request-Id:
              $ref: "#/components/headers/RequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "401":
          $ref: "#/components/responses/APIError"
        "404":
          $ref: "#/components/responses/APIError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Opaque workspace identifier. Clients must not infer semantics from it.
      schema:
        $ref: "#/components/schemas/ResourceId"
...
    ApplicationId:
      name: applicationId
      in: path
      required: true
      description: Opaque application identifier.
      schema:
        $ref: "#/components/schemas/ResourceId"
    ArtifactId:
      name: artifactId
      in: path
      required: true
      description: Opaque artifact identifier.
      schema:
        $ref: "#/components/schemas/ResourceId"
    ReleaseId:
      name: releaseId
      in: path
      required: true
      description: Opaque release identifier.
      schema:
        $ref: "#/components/schemas/ResourceId"
    PageCursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor returned by the previous page.
      schema:
        $ref: "#/components/schemas/Cursor"
    PageSize:
      name: limit
      in: query
      required: false
      description: Maximum number of records to return.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Caller-generated key used to replay a mutation safely.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$"
  schemas:
    ResourceId:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^[A-Za-z0-9][A-Za-z0-9_-]{0,127}$"
      description: Opaque resource identifier. Clients must compare it only as a string.
    Timestamp:
      type: string
      format: date-time
      description: UTC RFC 3339 timestamp. Writers emit a trailing `Z`.
      example: "2026-08-18T12:00:00Z"
    Cursor:
      type: string
      minLength: 1
      maxLength: 1024
      description: Opaque, endpoint-specific pagination cursor.
    PageInfo:
      type: object
      required: [hasMore]
      properties:
        nextCursor:
          allOf:
            - $ref: "#/components/schemas/Cursor"
          nullable: true
        hasMore:
          type: boolean
      additionalProperties: false
...
    ArtifactSource:
      type: object
      required: [repository, ref, sha]
      properties:
        repository:
          type: string
          format: uri
          maxLength: 2048
          description: HTTPS source repository URL without embedded credentials or query parameters.
        ref:
          type: string
          minLength: 1
          maxLength: 512
        sha:
          type: string
          pattern: "^(?:[a-fA-F0-9]{40}|[a-fA-F0-9]{64})$"
          description: Full immutable source commit identifier.
      additionalProperties: false
    ArtifactEvidence:
      type: object
      properties:
        sbom:
          type: string
          format: uri
          maxLength: 2048
        provenance:
          type: string
          format: uri
          maxLength: 2048
        scan:
          type: string
          format: uri
          maxLength: 2048
        ciRun:
          type: string
          format: uri
          maxLength: 2048
      additionalProperties: false
      description: Safe HTTPS metadata references. KubeOrch stores but does not fetch these URLs during registration.
    Artifact:
      type: object
      required: [id, workspaceId, image, digest, source, evidence, createdBy, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        workspaceId:
          $ref: "#/components/schemas/ResourceId"
        image:
          type: string
          maxLength: 512
          pattern: "^.+@sha256:[a-f0-9]{64}$"
          description: Canonical OCI repository and immutable digest. Mutable tags are not retained as identity.
        digest:
          type: string
          pattern: "^sha256:[a-f0-9]{64}$"
        source:
          $ref: "#/components/schemas/ArtifactSource"
        evidence:
          $ref: "#/components/schemas/ArtifactEvidence"
        createdBy:
          $ref: "#/components/schemas/ResourceId"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
      additionalProperties: false
    CreateArtifactRequest:
      type: object
      required: [image, source]
      properties:
        image:
          type: string
          maxLength: 512
          pattern: "^.+@sha256:[a-fA-F0-9]{64}$"
          description: OCI image reference pinned by digest. A tag without a digest is rejected.
        source:
          $ref: "#/components/schemas/ArtifactSource"
        evidence:
          $ref: "#/components/schemas/ArtifactEvidence"
      additionalProperties: false
    ArtifactList:
      type: object
      required: [items, pageInfo]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Artifact"
        pageInfo:
          $ref: "#/components/schemas/PageInfo"
      additionalProperties: false
    ReleaseSource:
      type: string
      enum: [external-ci, manual]
    Release:
      type: object
      required: [id, workspaceId, applicationId, applicationRevision, artifactIds, source, createdBy, createdAt]
      properties:
        id:
          $ref: "#/components/schemas/ResourceId"
        workspaceId:
          $ref: "#/components/schemas/ResourceId"
        applicationId:
          $ref: "#/components/schemas/ResourceId"
        applicationRevision:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:/@+\\-]{0,127}$"
        artifactIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ResourceId"
        source:
          $ref: "#/components/schemas/ReleaseSource"
        sourceReference:
          type: string
          format: uri
          maxLength: 2048
          description: Required HTTPS CI-run reference when source is external-ci.
        createdBy:
          $ref: "#/components/schemas/ResourceId"
        createdAt:
          $ref: "#/components/schemas/Timestamp"
      additionalProperties: false
    CreateReleaseRequest:
      type: object
      required: [applicationRevision, artifactIds, source]
      properties:
        applicationRevision:
          type: string
          minLength: 1
          maxLength: 128
          pattern: "^[A-Za-z0-9][A-Za-z0-9._:/@+\\-]{0,127}$"
          description: Immutable desired-state or canvas revision supplied by the caller.
        artifactIds:
          type: array
          minItems: 1
          maxItems: 100
          uniqueItems: true
          items:
            $ref: "#/components/schemas/ResourceId"
        source:
          $ref: "#/components/schemas/ReleaseSource"
        sourceReference:
          type: string
          format: uri
          maxLength: 2048
          description: Required HTTPS CI-run reference when source is external-ci.
      additionalProperties: false
    ReleaseList:
      type: object
      required: [items, pageInfo]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        pageInfo:
          $ref: "#/components/schemas/PageInfo"
      additionalProperties: false
    FieldViolation:
      type: object
      required: [field, code, message]
      properties:
        field:
          type: string
          description: JSON Pointer to the invalid field when available.
        code:
          type: string
          description: Stable machine-readable validation code.
        message:
          type: string
          description: Human-readable validation message.
      additionalProperties: false
    APIError:
      type: object
      required: [type, title, status, code, requestId]
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable problem type URI.
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        code:
          type: string
          description: Stable KubeOrch error code suitable for programmatic handling.
        requestId:
          type: string
          description: Safe correlation identifier; never contains credentials.
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldViolation"
      additionalProperties: false
...
  responses:
...
    APIError:
      description: Request failed using the Beta API error envelope.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/RequestId"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/APIError"
  headers:
    RequestId:
      description: Correlation identifier for support and audit records.
      schema:
        type: string
        minLength: 1
        maxLength: 128
    IdempotencyReplayed:
      description: "`true` when the response was replayed from an earlier idempotent request."
      schema:
        type: boolean